Moves tests to middleware/receive-message dir
