+++ lib/middleware/receive-message/map-request-params.js
@@ -29,9 +29,6 @@ function parseIsNewConversation(req) {
 
 module.exports = function mapRequestParams(config = {}) {
   return (req, res, next) => {
-    // TODO: Eventually deprecate client.
-    req.client = config.client;
-
     /**
      * config.paramsMap is a hash that holds the names of the new req properties that will
      * be parsed from the req.body object.
